+++ src/app/shared/directives/set-background.directive.ts
@@ -1,22 +1,32 @@
-import { Directive, ElementRef, HostListener, Input } from '@angular/core';
+import { Directive, ElementRef, HostListener, HostBinding, Input } from '@angular/core';
 
 @Directive({
   selector: '[appSetBackground]'
 })
 export class SetBackgroundDirective {
-  @Input('highlight') color: string;
+  @Input('appSetBackground') color: string;
 
-  constructor(private element: ElementRef) { }
+  @HostBinding('style.border') border: string;
 
-  @HostListener('mouseenter') onMouseEnter () {
+  private element: HTMLElement;
+
+  constructor(private el: ElementRef) {
+    this.element = el.nativeElement;
+   }
+
+  @HostListener('mouseenter')
+  onMouseEnter () {
     this.highlight(this.color || 'grey');
+    this.border = '1px solid darkgrey';
   }
-  @HostListener('mouseleave') onMouseLeave () {
+  @HostListener('mouseleave')
+  onMouseLeave () {
     this.highlight(null);
+    this.border = 'none';
   }
 
   private highlight(color: string) {
-    this.element.nativeElement.style.backgroundColor = color;
+    this.element.style.backgroundColor = color;
   }
 
 }